requirements.txt
pdfplumber>=0.10.3
python-dotenv>=1.0.1
weaviate-client>=4.4b9
